+++ Rlog/View/Components/InputFormElement.swift
@@ -1,5 +1,5 @@
 //
-//  CustomTextFieldContainer.swift
+//  InputFormElement.swift
 //  Rlog
 //
 //  Created by Noah's Ark on 2022/10/19.
@@ -8,10 +8,10 @@
 import SwiftUI
 
 struct InputFormElement: View {
-    let containerType: UnderlinedTextFieldType
+    let containerType: BorderedTextFieldType
     var text: Binding<String>
     
-    init(containerType: UnderlinedTextFieldType, text: Binding<String>) {
+    init(containerType: BorderedTextFieldType, text: Binding<String>) {
         self.containerType = containerType
         self.text = text
     }
@@ -48,11 +48,11 @@ private extension InputFormElement {
     
     var workplaceView: some View {
         VStack {
-            UnderlinedTextField(
+            BorderedTextField(
                 textFieldType: .workplace,
                 text: text
             )
-
+            
             if text.wrappedValue.count == 20 {
                 HStack {
                     Text("20자 이상 입력할 수 없어요.")
@@ -66,16 +66,10 @@ private extension InputFormElement {
     
     var wageView: some View {
         VStack {
-            HStack {
-                UnderlinedTextField(
-                    textFieldType: .wage,
-                    text: text
-                )
-                
-                Spacer()
-                
-                Text("원")
-            }
+            BorderedTextField(
+                textFieldType: .wage,
+                text: text
+            )
             
             if let textToInt = Int(text.wrappedValue), textToInt >= 1000000 {
                 HStack {
@@ -86,52 +80,44 @@ private extension InputFormElement {
                 }
                 .padding(.top, 4)
             }
-
         }
     }
     
     var paydayView: some View {
-        HStack(alignment: .top) {
-            Text("매월")
-            Spacer()
-            VStack(spacing: 0) {
-                UnderlinedTextField(
-                    textFieldType: .payday,
-                    text: text
-                )
-                
-                if let textToInt = Int16(text.wrappedValue), textToInt > 28 || textToInt < 1 {
-                    HStack {
-                        Text("1~28 사이의 숫자를 입력해주세요")
-                            .font(.footnote)
-                            .foregroundColor(.red)
-                        Spacer()
-                    }
-                    .padding(.top, 4)
-                }
-            }
-            Spacer()
+        VStack(spacing: 0) {
+            BorderedTextField(
+                textFieldType: .payday,
+                text: text
+            )
             
-            Text("일")
+            if let textToInt = Int16(text.wrappedValue), textToInt > 28 || textToInt < 1 {
+                HStack {
+                    Text("1~28 사이의 숫자를 입력해주세요")
+                        .font(.footnote)
+                        .foregroundColor(.red)
+                    Spacer()
+                }
+                .padding(.top, 4)
+            }
         }
     }
     
     var reasonView: some View {
-        UnderlinedTextField(
+        BorderedTextField(
             textFieldType: .reason,
             text: text
         )
     }
-
+    
     var timeView: some View {
-        UnderlinedTextField(
+        BorderedTextField(
             textFieldType: .time,
             text: text
         )
     }
     
     var noneView: some View {
-        UnderlinedTextField(
+        BorderedTextField(
             textFieldType: .none(title: containerType.title),
             text: text
         )
